refactor: move saving logic to one file
