+++ frontend/src/App.vue
@@ -13,7 +13,7 @@ import { onMounted, h, ref, provide } from 'vue'
 import { FrappeUIProvider, toast } from 'frappe-ui'
 
 import { Wifi, WifiOff } from 'lucide-vue-next'
-import { syncPresentationToServer } from '@/stores/presentation'
+import { syncPresentationToServer } from '@/utils/indexedDB'
 
 const isOnline = ref(false)
