+++ frontend/src/utils/indexedDB.js
@@ -1,3 +1,13 @@
+import { ref, computed } from 'vue'
+import {
+	presentationDoc,
+	presentationId,
+	hasStateChanged,
+	savePresentationDoc,
+} from '@/stores/presentation'
+import { slides } from '@/stores/slide'
+import { cloneObj } from '@/utils/helpers'
+
 const DB_NAME = 'slides-db'
 const DB_VERSION = 1
 const STORE = 'presentations'
@@ -76,4 +86,80 @@ const getPresentationFromLocalDB = async (id) => {
 	})
 }
 
-export { savePresentationToLocalDB, getPresentationFromLocalDB }
+const dirtySince = ref(null)
+
+const isDirty = computed(() => {
+	if (!presentationDoc.value || !slides.value) return false
+
+	const original = JSON.parse(JSON.stringify(presentationDoc.value.slides || []))
+	const current = JSON.parse(JSON.stringify(slides.value || []))
+
+	return hasStateChanged(original, current)
+})
+
+const isSaving = ref(false)
+
+let syncThumbnail = 0
+const syncOfflineChangesStatus = ref(null)
+
+const syncPresentationToServer = async (hadDroppedConnection) => {
+	isSaving.value = true
+
+	try {
+		const snapshot = await getPresentationFromLocalDB(presentationId.value)
+
+		if (!snapshot || !snapshot.dirty) return
+
+		if (hadDroppedConnection) {
+			syncOfflineChangesStatus.value = 'Syncing local changes...'
+		}
+
+		await savePresentationDoc(snapshot.content)
+
+		if (hadDroppedConnection) {
+			syncOfflineChangesStatus.value = 'All changes synced'
+			setTimeout(() => {
+				syncOfflineChangesStatus.value = null
+			}, 2000)
+		}
+
+		await savePresentationToLocalDB({
+			...snapshot,
+			dirty: false,
+			updatedAt: Date.now(),
+		})
+	} catch (err) {
+		console.error('Sync to server failed: ', err)
+	} finally {
+		isSaving.value = false
+	}
+}
+
+const getLatestSlideContent = () => {
+	const latestContent = slides.value
+	return cloneObj(latestContent)
+}
+
+const saveChanges = async () => {
+	if (isSaving.value) return
+
+	if (!isDirty.value && syncThumbnail === 0) return
+
+	if (isDirty.value) syncThumbnail = 1
+	else syncThumbnail = 0
+
+	const content = getLatestSlideContent()
+
+	await savePresentationToLocalDB({
+		id: presentationId.value,
+		content: content,
+		updatedAt: Date.now(),
+		dirty: true,
+	})
+
+	if (!navigator.onLine) return
+
+	await syncPresentationToServer()
+}
+
+export { syncPresentationToServer, saveChanges, dirtySince, isDirty, syncOfflineChangesStatus }
